% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals_function.R
\name{cal_intervals}
\alias{cal_intervals}
\title{Calculating Intervals}
\usage{
cal_intervals(timetable, breaks = "1 min")
}
\arguments{
\item{timetable}{dataframe containing minimum two columns indicating starting point and end point of each recording (each row is one recording); columns need to be called start and end for the designated timestamps}

\item{breaks}{Duration between the intervals given as a String. Defaults to 1 min}
}
\description{
This function calculates intervals for a given break-duration between a given start time (start) and end time (end) for multiple recordings. Returns only the intervals as a vector to append to existing dataframe.
}
\keyword{HK}
\keyword{bats}
\keyword{intervals}
