% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals_function.R
\name{count_contacts_per_interval_of_sp}
\alias{count_contacts_per_interval_of_sp}
\title{Counting contacts for given Species}
\usage{
count_contacts_per_interval_of_sp(
  intervals,
  rufe_HKX,
  art,
  interval_length = 60
)
}
\arguments{
\item{intervals}{dataframe containing a column indicating the start point of each interval}

\item{rufe_HKX}{dataframe containing a column with time stamps (rz) for each call recorded as well a a column with species identity (Art)}

\item{art}{String of the species that the calls should be filtered by}

\item{interval_length}{Length of the interval in seconds. Defaults to 60s}
}
\description{
This function counts the number of contacts across set intervals provided filtering beforehand by a single species.
Results are returned as a vector suited for appending to original interval dataframe
}
\examples{
cat_function()
}
\keyword{HK}
\keyword{bats}
