% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals-package.R
\docType{package}
\name{intervals-package}
\alias{intervals}
\alias{intervals-package}
\title{intervals: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ds-turner/intervals}
  \item Report bugs at \url{https://github.com/ds-turner/intervals/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
