% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ints-package.R
\docType{package}
\name{ints-package}
\alias{ints}
\alias{ints-package}
\title{ints: What the Package Does (One Line, Title Case)}
\description{
Functions for handling ints. The primary functions are `grp_ints` and `merge_ints`, which group overlapping or continuous ints and merge the overlapping ints, respectively.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ds-turner/ints}
  \item \url{https://ds-turner.github.io/ints/}
  \item Report bugs at \url{https://github.com/ds-turner/ints/issues}
}

}
\author{
\strong{Maintainer}: David Turner \email{david.samuel.turner@gmail.com.com}

}
\keyword{internal}
