% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_ints.R
\name{grp_ints}
\alias{grp_ints}
\title{Group Overlapping Intervals}
\usage{
grp_ints(.data, .start, .end, ..., .gap = 1, .group_col = int_grp_id)
}
\arguments{
\item{.data}{A data frame containing the interval data.}

\item{.start}{The column name (unquoted) representing the start of the intervals.}

\item{.end}{The column name (unquoted) representing the end of the intervals.}

\item{...}{Additional columns to group by before identifying overlapping intervals.}

\item{.gap}{The maximum allowed gap between intervals for them to be considered overlapping.
Intervals are grouped if the start of one interval is less than or equal to
the end of the previous interval plus the gap. Default is 1.}

\item{.group_col}{The name of the column that will indicate the group ID for overlapping intervals.  Default is \code{int_grp_id}.}
}
\value{
A \code{data.table} with the original data and an additional column indicating
the group ID for overlapping intervals (the \code{.group_id} column).
}
\description{
This function groups overlapping intervals based on its start and end.
It assigns a unique group ID to each set of overlapping intervals.
Intervals are considered overlapping if the start of one interval is within the gap
threshold of the end of another interval.
}
\examples{
\dontrun{
data <- data.frame(
  id = 1:5,
  start = c(1, 2, 5, 10, 12),
  end = c(3, 4, 7, 11, 14)
)

# Group intervals with a gap of 1
grp_ints(data, start, end, .gap = 1)
}

}
