% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals-package.R
\docType{package}
\name{intervals-package}
\alias{intervals}
\alias{intervals-package}
\title{intervals: What the Package Does (One Line, Title Case)}
\description{
Functions for handling intervals. The primary functions are `grp_ints` and `merge_ints`, which group overlapping or continuous intervals and merge the overlapping intervals, respectively.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ds-turner/intervals}
  \item Report bugs at \url{https://github.com/ds-turner/intervals/issues}
}

}
\author{
\strong{Maintainer}: David Turner \email{david.samuel.turner@gmail.com.com}

}
\keyword{internal}
